requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
binaryornot==0.4.4
celery==5.5.3
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cookiecutter==2.6.0
cron-descriptor==1.4.5
cryptography==43.0.1
daphne==4.1.2
dj-database-url==3.0.0
Django==5.1.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.4
kombu==5.5.4
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
msgpack==1.1.0
packaging==25.0
pillow==11.0.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
rich==13.9.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
Twisted==24.7.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
zope.interface==7.1.0
